Extract Deamon monad code into module
We introduce a new `Radicle.Daemon.Monad` module where we move the code
that defines the `Daemon` monad, the environment, and the errors.

This module also provides the new `liftExcept` function that reduces
code duplication.
